requirements.txt
beautifulsoup4
requests
gradio
pillow